+++ src/main/java/com/taskmanager/service/impl/UserServiceImpl.java
@@ -8,11 +8,15 @@ import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
 import com.taskmanager.dto.MyUserDto;
 import com.taskmanager.dto.MyUserResponse;
+import com.taskmanager.exceptions.ActiveUserCannotBeDeletedException;
 import com.taskmanager.exceptions.MyUserNotFoundException;
 import com.taskmanager.model.MyUser;
 import com.taskmanager.repository.MyUserRepository;
@@ -85,14 +89,18 @@ public class UserServiceImpl implements MyUserService {
 		return myUser;
 	}
 
-	public MyUser createUserUpdate(MyUser myUserDto, MyUserDto myUserDtoUpdate) {
+	public MyUser createUserUpdate(MyUser myUser, MyUserDto myUserDtoUpdate) {
+
 		logger.trace("Entered......createUserUpdate() ");
 
-		myUserDto.setUsername(myUserDtoUpdate.getUsername());
-		myUserDto.setUsername(myUserDtoUpdate.getUsername());
-		myUserDto.setRole(myUserDtoUpdate.getRole());
+		myUser.setId(myUserDtoUpdate.getId());
+		myUser.setUsername(myUserDtoUpdate.getUsername());
+		myUser.setUsername(myUserDtoUpdate.getUsername());
+		myUser.setRole(myUserDtoUpdate.getRole());
+
 		logger.trace("Exited......createUserUpdate() ");
-		return myUserDto;
+
+		return myUser;
 	}
 
 	public List<MyUserDto> findAllUsers() {
@@ -132,18 +140,23 @@ public class UserServiceImpl implements MyUserService {
 	public MyUserDto getMyUserById(int id) {
 		logger.trace("Entered...........................getMyUserById()");
 
-		MyUser myUserDto = myUserRepository.findById(id)
+		MyUser myUser = myUserRepository.findById(id)
 				.orElseThrow(() -> new MyUserNotFoundException("User Id could not be found :("));
 		logger.trace("Exited...........................getMyUserById()");
-		return mapToDto(myUserDto);
+		return mapToDto(myUser);
 	}
 
-	private MyUserDto mapToDto(MyUser myUserDto) {
+	private MyUserDto mapToDto(MyUser myUser) {
 		logger.trace("Entered...........................mapToDto()");
 
-		MyUserDto MyUserDto = new MyUserDto();
+		MyUserDto newMyUserDto = new MyUserDto();
+		newMyUserDto.setId(myUser.getId());
+		newMyUserDto.setPassword(myUser.getPassword());
+		newMyUserDto.setRole(myUser.getRole());
+		newMyUserDto.setUsername(myUser.getUsername());
+
 		logger.trace("Exited...........................mapToDto()");
-		return MyUserDto;
+		return newMyUserDto;
 	}
 
 	@Override
@@ -152,11 +165,11 @@ public class UserServiceImpl implements MyUserService {
 
 		try {
 			// Find the MyUser entity by ID or throw an exception if not found
-			MyUser myUserDto = myUserRepository.findById(id)
-					.orElseThrow(() -> new MyUserNotFoundException("MyUser could not be updated..."));
+			MyUser myUser = myUserRepository.findById(id)
+					.orElseThrow(() -> new MyUserNotFoundException("MyUser could not be found..."));
 
 			// Create an updated MyUser entity
-			MyUser updatedMyUser = createUserUpdate(myUserDto, myUserDtoUpdate);
+			MyUser updatedMyUser = createUserUpdate(myUser, myUserDtoUpdate);
 
 			// Save the updated MyUser entity
 			MyUser newMyUser = myUserRepository.save(updatedMyUser);
@@ -176,10 +189,43 @@ public class UserServiceImpl implements MyUserService {
 	public void deleteMyUserById(int id) {
 		logger.trace("Entered......deleteMyUserById() ");
 
+		Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+
+		UserDetails userDetails = (UserDetails) authentication.getPrincipal();
+
+		String activeUsername = userDetails.getUsername();
+
 		MyUser myUser = myUserRepository.findById(id)
 				.orElseThrow(() -> new MyUserNotFoundException("MyUser could not be deleted..."));
+
+		if (myUserRepository.findById(id).isPresent()) {
+
+			if (myUser.getUsername().equals(activeUsername)) {
+				throw new ActiveUserCannotBeDeletedException("Active user cannot be deleted...");
+			} else {
+
+				myUserRepository.deleteById(id);
+			}
+		}
+
 		logger.trace("Exited......deleteMyUserById() ");
-		myUserRepository.deleteById(id);
+	}
+
+	@Override
+	public MyUserDto currentUser() {
+
+		Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+
+		System.out.println("authentication.getPrincipal() = " + authentication.getPrincipal());
+
+		UserDetails userDetails = (UserDetails) authentication.getPrincipal();
+		String username = userDetails.getUsername();
+		System.out.println("username ================= " + username);
+
+		MyUser myUser = myUserRepository.findByUsername(userDetails.getUsername())
+				.orElseThrow(() -> new MyUserNotFoundException("MyUser could not be found..."));
+
+		return mapToDto(myUser);
 	}
 
 }
